Tidy up ProjectProtoTransformProvider.
Since it's not safe for implementations of ProjectProtoTransform to
use many parts of QuerySyncProject, remove it from the factory method.
Replace it with ProjectDefinition which is safe to use here, and is
needed by real implementions.

Update AndroidProjectProtoTransform accordingly.

Bug: 357876275
Test: no functional changes in this CL.
Change-Id: I5015d976fa735b48b6ea64078f57babeee4d1f73

AOSP: b1fe2e07d2847b8d5b9b6072828a686763196cf3