Improve artifact metadata handling and storage.
Store the metadata inside the BuildArtifact class itself. Update the
proto definitions to reflect this.

We add specific classes for storing each type of metadata, and also to
convert this to proto. We also have specific interfaces for extracting
metadata from build artifacts, and a factory for converting a proto back
to a java class.

The new metadata support is somewhat generic and extensible: the main
place this falls down is in the proto definition where all types of
metadata are handled by a single proto message.

Bug: 357876275
Test: Unit tests updated accordingly.

Change-Id: I88262c057e6260f80cd32f04cd589c9f99911d25

AOSP: 46d1745775b1b007472aa3a23e14dddb0a0eafc9